requirements.txt
camelot-py==0.4.0
chardet==3.0.4
Click==7.0
cycler==0.10.0
distro==1.3.0
et-xmlfile==1.0.1
flake8==3.6.0
jdcal==1.4
kiwisolver==1.0.1
matplotlib==3.0.2
mccabe==0.6.1
numpy==1.15.4
opencv-python==3.4.4.19
openpyxl==2.5.12
pandas==0.23.4
pdfminer.six==20181108
pycodestyle==2.4.0
pycryptodome==3.7.2
pyflakes==2.0.0
pyparsing==2.3.0
PyPDF2==1.26.0
python-dateutil==2.7.5
pytz==2018.7
selenium==3.141.0
six==1.11.0
sortedcontainers==2.1.0
tabula-py==1.3.1
urllib3==1.24.1
